+++ zest/releaser/lasttagdiff.py
@@ -12,8 +12,7 @@ logger = logging.getLogger(__name__)
 
 
 def main():
-    logging.basicConfig(level=utils.loglevel(),
-                        format="%(levelname)s: %(message)s")
+    utils.configure_logging()
     vcs = zest.releaser.choose.version_control()
     if len(sys.argv) > 1:
         found = sys.argv[-1]
@@ -26,5 +25,5 @@ def main():
     logger.info("Showing differences from the last commit against tag %s",
                 full_tag)
     diff_command = vcs.cmd_diff_last_commit_against_tag(found)
-    print diff_command
-    print system(diff_command)
+    print(diff_command)
+    print(system(diff_command))