+++ zest/releaser/lasttaglog.py
@@ -12,8 +12,7 @@ logger = logging.getLogger(__name__)
 
 
 def main():
-    logging.basicConfig(level=utils.loglevel(),
-                        format="%(levelname)s: %(message)s")
+    utils.configure_logging()
     vcs = zest.releaser.choose.version_control()
     if len(sys.argv) > 1:
         found = sys.argv[-1]
@@ -26,5 +25,5 @@ def main():
     logger.info("Showing log since tag %s and the last commit.",
                 full_tag)
     log_command = vcs.cmd_log_since_tag(found)
-    print log_command
-    print system(log_command)
+    print(log_command)
+    print(system(log_command))