+++ zest/releaser/prerelease.py
@@ -166,8 +166,8 @@ class Prereleaser(baserelease.Basereleaser):
             # python2.6.2 bug... http://bugs.python.org/issue5170 This is the
             # spot it can surface as we show a part of the changelog which can
             # contain every kind of character.  The rest is mostly ascii.
-            print "Diff results:"
-            print diff
+            print("Diff results:")
+            print(diff)
         else:
             # Common case
             logger.info("The '%s':\n\n%s\n" % (diff_cmd, diff))
@@ -186,7 +186,6 @@ def datacheck(data):
 
 def main():
     utils.parse_options()
-    logging.basicConfig(level=utils.loglevel(),
-                        format="%(levelname)s: %(message)s")
+    utils.configure_logging()
     prereleaser = Prereleaser()
     prereleaser.run()