+++ zest/releaser/tests/test_setup.py
@@ -2,6 +2,7 @@ import re
 import tempfile
 
 import z3c.testsetup
+from colorama import Fore
 from zope.testing import renormalizing
 
 
@@ -54,6 +55,10 @@ checker = renormalizing.RENormalizing([
      'On branch'),
     # Hg 3.3 prints 'committing files' or 'committing <filename>'
     (re.compile(r'^committing.*'), ''),
+    # We should ignore coloring by colorama.  This catches Fore.RED.
+    # Note the extra backslash in front of the left bracket, otherwise
+    # you get: "error: unexpected end of regular expression"
+    (re.compile(r'\x1b\[31m'), ''),
 ])