+++ zest/releaser/utils.py
@@ -1,4 +1,5 @@
 # Small utility methods.
+from colorama import Fore
 from optparse import OptionParser
 from pkg_resources import parse_version
 import logging
@@ -84,12 +85,12 @@ def get_input(question):
         # Normal operation.
         return raw_input(question).strip()
     # Testing means no interactive input. Get it from answers_for_testing.
-    print "Question:", question
+    print("Question: %s" % question)
     answer = test_answer_book.get_next_answer()
     if answer == '':
-        print "Our reply: <ENTER>"
+        print("Our reply: <ENTER>")
     else:
-        print "Our reply:", answer
+        print("Our reply: %s" % answer)
     return answer
 
 
@@ -155,8 +156,8 @@ def ask(question, default=True, exact=False):
         if not answer and default is not None:
             return default
         if exact and answer.lower() not in ('yes', 'no'):
-            print ("Please explicitly answer yes/no in full "
-                   "(or accept the default)")
+            print("Please explicitly answer yes/no in full "
+                  "(or accept the default)")
             continue
         if answer:
             answer = answer[0].lower()
@@ -165,7 +166,7 @@ def ask(question, default=True, exact=False):
             if answer == 'n':
                 return False
         # We really want an answer.
-        print 'Please explicitly answer y/n'
+        print('Please explicitly answer y/n')
         continue
 
 
@@ -235,28 +236,35 @@ def extract_headings_from_history(history_lines):
     return headings
 
 
-def show_first_and_last_lines(result):
-    """Just print the first and last five lines of (pypi) output"""
+def show_interesting_lines(result):
+    """Just print the first and last five lines of (pypi) output.
+
+    But: when there are errors or warnings, print everything and ask
+    the user if she wants to continue.
+    """
+    if Fore.RED in result:
+        # warnings/errors, print complete result.
+        print(result)
+        if not ask(
+                "There were errors or warnings. Are you sure "
+                "you want to continue?", default=False):
+            sys.exit(1)
+        # User has seen everything and wants to continue.
+        return
+
+    # No errors or warnings.  Show first and last lines.
     lines = [line for line in result.split('\n')]
     if len(lines) < 11:
         for line in lines:
-            print line
+            print(line)
         return
-    print 'Showing first few lines...'
+    print('Showing first few lines...')
     for line in lines[:5]:
-        print line
-    print '...'
-    print 'Showing last few lines...'
+        print(line)
+    print('...')
+    print('Showing last few lines...')
     for line in lines[-5:]:
-        print line
-
-
-def show_last_lines(result):
-    """Just print the last five lines of (pypi) output"""
-    lines = [line for line in result.split('\n')]
-    print 'Showing last few lines...'
-    for line in lines[-5:]:
-        print line
+        print(line)
 
 
 def setup_py(rest_of_cmdline):
@@ -275,11 +283,11 @@ def is_data_documented(data, documentation={}):
     """check that the self.data dict is fully documented"""
     if TESTMODE:
         # Hack for testing to prove entry point is being called.
-        print "Checking data dict"
+        print("Checking data dict")
     undocumented = [key for key in data
                     if key not in documentation]
     if undocumented:
-        print 'Internal detail: key(s) %s are not documented' % undocumented
+        print('Internal detail: key(s) %s are not documented' % undocumented)
 
 
 def resolve_name(name):
@@ -429,7 +437,7 @@ def prepare_documentation_entrypoint(data):
             result.append('')
 
     open(target, 'wb').write('\n'.join(result))
-    print "Wrote entry point documentation to", target
+    print("Wrote entry point documentation to %s" % target)
 
 
 def system(command, input=''):
@@ -437,7 +445,12 @@ def system(command, input=''):
     logger.debug("Running command: %r", command)
     if command.startswith(sys.executable):
         env = dict(os.environ, PYTHONPATH=os.pathsep.join(sys.path))
-        show_stderr = False
+        if ' upload ' in command or ' register ' in command:
+            # We really do want to see the stderr here, otherwise a
+            # failed upload does not even show up in the output.
+            show_stderr = True
+        else:
+            show_stderr = False
     else:
         env = None
         show_stderr = True
@@ -454,9 +467,20 @@ def system(command, input=''):
     i.close()
     stdout_output = o.read()
     stderr_output = e.read()
+    # TODO.  Note that the returncode is always None, also after
+    # running p.kill().  The shell=True may be tripping us up.  For
+    # some ideas, see http://stackoverflow.com/questions/4789837
     if p.returncode or show_stderr or 'Traceback' in stderr_output:
         # Some error occured
-        result = stdout_output + stderr_output
+        # print(Fore.RED + stderr_output)
+        stderr_output = stderr_output.strip()
+        if stderr_output:
+            # Make sure every error line is marked red.
+            errors = [(Fore.RED + line) for line in stderr_output.split('\n')]
+            errors = '\n'.join(errors)
+        else:
+            errors = ''
+        result = stdout_output + errors
     else:
         # Only return the stdout. Stderr only contains possible
         # weird/confusing warnings that might trip up extraction of version
@@ -554,5 +578,12 @@ You may want to quit and fix this.
         q += "Do you want to continue with the release?"
         if not ask(q, default=False):
             return False
-        print q
+        print(q)
     return True
+
+
+def configure_logging():
+    logging.addLevelName(logging.WARNING, Fore.MAGENTA + logging.getLevelName(logging.WARNING))
+    logging.addLevelName(logging.ERROR, Fore.RED + logging.getLevelName(logging.ERROR))
+    logging.basicConfig(level=loglevel(),
+                        format="%(levelname)s: %(message)s")