+++ zest/releaser/vcs.py
@@ -45,11 +45,11 @@ class BaseVersionControl(object):
             # UserWarnings.
             system(utils.setup_py('egg_info'))
             version = system(utils.setup_py('--version')).splitlines()[0]
-            if version.startswith('Traceback'):
+            if 'Traceback' in version:
                 # Likely cause is for example forgetting to 'import
                 # os' when using 'os' in setup.py.
                 logger.critical('The setup.py of this package has an error:')
-                print version
+                print(version)
                 logger.critical('No version found.')
                 sys.exit(1)
             return utils.strip_version(version)
@@ -286,7 +286,7 @@ class BaseVersionControl(object):
         tagdir = self.prepare_checkout_dir(prefix)
         os.chdir(tagdir)
         cmd = self.cmd_checkout_from_tag(version, tagdir)
-        print system(cmd)
+        print(system(cmd))
 
     def is_clean_checkout(self):
         "Is this a clean checkout?"